- remove transcoding service - add getFileData readerService - use ffmpegService directly for transcoding
